Renamed slice_replacers.py
