+++ model-optimizer/extensions/front/onnx/AttributedSliceToSlice.py
